refactor: move workspace selection
